refactor: mark storage_location.go as pure functions compute storage location
