+++ core/state/storage_location.go
@@ -53,3 +53,29 @@ func StorageLocationOfFixedArrayElement(arraySlot StorageLocation, elementIndex 
 	)
 	return StorageLocation(new(big.Int).Add(arraySlot.Big(), offset).Bytes())
 }
+
+func GetValidatorOwnerSlot(candidate common.Address) common.Hash {
+	validatorMappingSlot := vicValidatorStorageMap["validatorsState"]
+	return crypto.Keccak256Hash(candidate.Hash().Bytes(), common.BigToHash(big.NewInt(int64(validatorMappingSlot))).Bytes())
+}
+
+func GetStorageKeyForSlot(slot uint64) common.Hash {
+	slotHash := common.BigToHash(new(big.Int).SetUint64(slot))
+	return slotHash
+}
+
+func GetStorageKeyForMapping(key common.Hash, slot uint64) common.Hash {
+	slotHash := common.BigToHash(new(big.Int).SetUint64(slot))
+	retByte := crypto.Keccak256(key.Bytes(), slotHash.Bytes())
+	ret := new(big.Int)
+	ret.SetBytes(retByte)
+	return common.BigToHash(ret)
+}
+
+// GetLocDynamicArrAtElement is used to get the location of element inside dynamic array
+func GetLocDynamicArrAtElement(slotHash common.Hash, index uint64, elementSize uint64) common.Hash {
+	slotKecBig := crypto.Keccak256Hash(slotHash.Bytes()).Big()
+	// arrBig = slotKecBig + index * elementSize
+	arrBig := slotKecBig.Add(slotKecBig, new(big.Int).SetUint64(index*elementSize))
+	return common.BigToHash(arrBig)
+}
